requirements.txt
boto3==1.17.105
Django==3.1.6
django-filter==2.4.0
djangorestframework
djangorestframework-simplejwt
django-storages==1.11.1
django-suit @ https://github.com/darklow/django-suit/tarball/v2
Pillow==7.0.0
PyJWT
simplejson==3.16.0
twilio
whitenoise==5.2.0
gunicorn
psycopg2-binary
dj-database-url
